Renamed ses -> cq
