+++ cq/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'cq.contrib.django.apps.SimpleEventSourcingApp'
