+++ cq/contrib/django/shortcuts.py
