+++ cq/exceptions.py
