+++ examples/djangoapp/settings_test.py
@@ -8,8 +8,8 @@ SECRET_KEY = ''.join([random.choice(string.ascii_letters) for x in range(40)])
 
 
 INSTALLED_APPS = (
-    'ses.contrib.django.apps.SimpleEventSourcingApp',
-    # 'ses',
+    'cq.contrib.django.apps.SimpleEventSourcingApp',
+    # 'cq',
     'accounts',
 )
