--- ses/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'ses.contrib.django.apps.SimpleEventSourcingApp'
