+++ tox.ini
@@ -20,7 +20,7 @@ changedir =
     d109: {toxinidir}/examples/djangoapp
     d110: {toxinidir}/examples/djangoapp
     sqla: {toxinidir}/examples/sqlalchemyapp
-    plain: {toxinidir}/ses
+    plain: {toxinidir}/cq
 passenv =
     DJANGO_SETTINGS_MODULE
 basepython =
